move test files to test folder
