+++ desktop.ini
@@ -1,0 +1,6 @@
+[.ShellClassInfo]
+IconResource=C:\windows\System32\SHELL32.dll,296
+[ViewState]
+Mode=
+Vid=
+FolderType=Generic
